% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asis.R
\name{as.data.frame.tb}
\alias{as.data.frame.tb}
\title{convert texblock to data.frame}
\usage{
\method{as.data.frame}{tb}(x, ...)
}
\arguments{
\item{x}{texblock object}

\item{\dots}{pass convert as a boolean argument to apply type.convert to the output columns}
}
\value{
data.frame
}
\description{
convert a texblock class into a data.frame
}
\examples{
x <- '1'
class(x) <- 'tb'
x1 <- x+x
as.data.frame(x1)
as.data.frame(x1,convert = TRUE)

}
