% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{as.tb}
\alias{as.tb}
\title{Try to coerce an object into a texblock object}
\usage{
as.tb(x)
}
\arguments{
\item{x}{object}
}
\value{
an object of class tb
}
\description{
coerce objects into a tb class object
}
