% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{crop}
\alias{crop}
\title{Crop texblocks}
\usage{
crop(x, i, j)
}
\arguments{
\item{x}{tb}

\item{i}{numeric, row indicies to extract}

\item{j}{numeric, column indicies to extract}
}
\value{
tb
}
\description{
Extract a subset of a texblock
}
\details{
idicies do not have to be consecutive, and can be -(...)
to extract all but the subset.
}
