% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_column.R, R/multi_row.R
\name{find_multicol}
\alias{find_multicol}
\alias{find_multirow}
\title{Find multirow or multicolumn}
\usage{
find_multicol(x)

find_multirow(x)
}
\arguments{
\item{x}{texblock}
}
\value{
data.frame
}
\description{
Returns indicies of multirow or multicolumn in a texblocks
}
\examples{
x <- multicol('a',2)
y <- multirow('a',2)
x\%>\%find_multicol()
y\%>\%find_multirow()

}
