% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{harvest}
\alias{harvest}
\title{Extract subsets of texblocks}
\usage{
harvest(x, I, J)
}
\arguments{
\item{x}{tb}

\item{I}{list, row indicies to extract}

\item{J}{list, column indicies to extract}
}
\value{
list of tb elements
}
\description{
Wrapper to crop that applies lists of indicies
}
\seealso{
\code{\link[purrr]{map2}}
}
