% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asis.R
\name{is.tb}
\alias{is.tb}
\title{Is the object of class tb}
\usage{
is.tb(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical - is the object of class tb
}
\description{
Is the object of class tb Very basic for many functions
in the package.
}
