% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{multirow}
\alias{multirow}
\alias{multicol}
\title{Merge texblocks}
\usage{
multirow(content, nrow, width = "*")

multicol(content, ncol, align = "c")
}
\arguments{
\item{content}{character, contents to display in the merged cell}

\item{nrow}{numeric, number of rows to span in the multirow}

\item{width}{character, width of the new element, Default: '*'}

\item{ncol}{numeric, number of columns to span in the multicol}

\item{align}{character, alignment of the merged column, Default: 'c'}
}
\value{
tb
}
\description{
Create multirow/multicol to texblock object
}
\details{
align is either l, c, r, or to have text wrapping specify a width p{5.0cm}.
\itemize{
\item for the width means the content's natural width.
}
}
\examples{
multirow('aaa',2)
multicol('aaa',2)
}
