% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_col.R
\name{pad}
\alias{pad}
\title{pad texblock}
\usage{
pad(x, times, sides)
}
\arguments{
\item{x}{tb}

\item{times}{numeric, number of columns to pad, Default: 1}

\item{sides}{character, which side to pad t,b,l or r}
}
\value{
tb
}
\description{
add columns/rows to texblock
}
