% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_col.R
\name{pad_col}
\alias{pad_col}
\title{pad columns of texblock}
\usage{
pad_col(x, cols = 1, side = 1)
}
\arguments{
\item{x}{tb}

\item{cols}{numeric, number of columns to pad, Default: 1}

\item{side}{numeric, which side to pad, 1='left' 2='right', Default: 1}
}
\value{
tb
}
\description{
add columns to texblock to the left
}
