% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular.R
\name{tabular}
\alias{tabular}
\title{TeX tabular object}
\usage{
tabular(x, align = NULL)
}
\arguments{
\item{x}{texblock object}

\item{align}{character, tabular alignment, Default: NULL}
}
\value{
texblock object
}
\description{
Wrap tabular syntax around a texblock object
}
