% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{tb_hline}
\alias{tb_hline}
\title{add hlines}
\usage{
tb_hline(x, lines = NULL)
}
\arguments{
\item{x}{texblock}

\item{lines}{lines to add hline, Default: NULL}
}
\value{
texblock
}
\description{
add hlines to texblock object
}
\details{
if NULL then all lines have hline added to them
}
