% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_col.R
\name{tb_reduce}
\alias{tb_reduce}
\title{reduce dimension of list of texblocks}
\usage{
tb_reduce(..., margin = 1, merge = FALSE)
}
\arguments{
\item{\dots}{values to reduce}

\item{margin}{numeric, how to reduce 1 by rows, 2 by columns, Default: 1}

\item{merge}{boolean, merge duplicated values into a multirow/multicolumn, Default: FALSE}
}
\value{
tb
}
\description{
reduce a list of texblock by rows or columns
}
