% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.tb}
\alias{as.data.frame.tb}
\title{convert texblock to data.frame}
\usage{
\method{as.data.frame}{tb}(x, ...)
}
\arguments{
\item{x}{texblock object}

\item{\dots}{not used}
}
\value{
data.frame
}
\description{
convert a texblock class into a data.frame
}
\examples{
x <- 'a'
class(x) <- 'tb'
x1 <- x+x
as.data.frame(x1)

}
