% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_col.R
\name{tb_reduce}
\alias{tb_reduce}
\title{reduce dimension of list of texblocks}
\usage{
tb_reduce(..., margin = c(1, 2))
}
\arguments{
\item{margin}{how to reduce 1 by rows, 2 by columns, Default: 1}

\item{x}{tb}
}
\value{
tb
}
\description{
reduce a list of texblock by rows or columns
}
