% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operators}
\alias{operators}
\alias{+.tb}
\alias{/.tb}
\title{texblock operators}
\usage{
\method{+}{tb}(e1, e2)

\method{/}{tb}(e1, e2)
}
\arguments{
\item{e1}{texblock object, right hand side object}

\item{e2}{texblock object, left hand side object}
}
\value{
texblock object
}
\description{
combine texblock object into bigger texblock objects
}
\details{
\itemize{
\item : joins e1 to e2 horizontally
}

/ : joins e1 to e2 vertically with no hline betweet them
\itemize{
\item : joins e1 to e2 vertically with an hline between them
}
}
\examples{

x <- '$\\\\alpha$'
y <- 'cindy'
class(x) <- 'tb'
class(y) <- 'tb'
x1 <- x+y+x
x1
x1 + x1
x1 / x1
x2 <- x1 / x1
x2 + x2
x3 <- x2/x2
x2 + x3

}
